% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{disaggregate}
\alias{disaggregate}
\title{disaggregate data}
\usage{
disaggregate(dat, counts)
}
\arguments{
\item{dat}{a data.frame}

\item{counts}{frequency counts for each row}
}
\description{
disaggregate data
}
