% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchShinyPopSize}
\alias{launchShinyPopSize}
\title{Launches the Shiny Application for Population Size}
\usage{
launchShinyPopSize(app = c("estimation", "power"))
}
\arguments{
\item{app}{Which application to launch.}
}
\description{
Launches the Shiny Application for Population Size
}
\details{
The manual for this shiny application is located at https://fellstat.github.io/shinyrecap/
}
