% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lcmcrSample}
\alias{lcmcrSample}
\title{Perform LCMCR sampling with a monitor function}
\usage{
lcmcrSample(
  object,
  burnin = 10000,
  samples = 1000,
  thinning = 10,
  clear_buffer = FALSE,
  output = TRUE,
  nMonitorBreaks = 100,
  monitorFunc = function(subs, tot) {
 }
)
}
\arguments{
\item{object}{the samples}

\item{burnin}{MCMC burn in}

\item{samples}{number of samples}

\item{thinning}{MCMC thinning}

\item{clear_buffer}{buffer clear buffer of object}

\item{output}{output progress}

\item{nMonitorBreaks}{number of times to call the monitor function}

\item{monitorFunc}{A function called nMonitorBreaks times taking the number of samples to be taken, and the total samples}
}
\description{
Perform LCMCR sampling with a monitor function
}
\details{
An edited version of \code{lcmCR_PostSampl}
}
