% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{simulateCapture}
\alias{simulateCapture}
\title{Simulate Capture Re-capture with heterogeneity}
\usage{
simulateCapture(hetero, p)
}
\arguments{
\item{hetero}{The heterogineity}

\item{p}{A vector of capture event probabilities}
}
\description{
Simulate Capture Re-capture with heterogeneity
}
\examples{
het <- simulateHeteroNormal(1000, 1.1)
cap <- simulateCapture(het, p = c(.05,.1,.05,.1))
summary(cap)
}
