% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{simulateHeteroNormal}
\alias{simulateHeteroNormal}
\title{simulate capture heterogineity}
\usage{
simulateHeteroNormal(N, heteroPerc = 1, perc = 0.9)
}
\arguments{
\item{N}{Population size}

\item{heteroPerc}{The increase in odds of capture for the perc 90th percentile most likely to be captured individuals, compared to the average individual.}

\item{perc}{The percentile to use.}
}
\description{
simulate capture heterogineity
}
\examples{
het <- simulateHeteroNormal(100, 1.1)
hist(het)
}
