% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{srhelp}
\alias{srhelp}
\title{A simple wrapper for shinyhelper::helper}
\usage{
srhelp(x, content, ...)
}
\arguments{
\item{x}{The shiny object to decorate}

\item{content}{the name of the help}

\item{...}{additional parameters for shinyhelper::helper}
}
\description{
A simple wrapper for shinyhelper::helper
}
