% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{shinyrecap-package}
\alias{shinyrecap-package}
\title{Shiny User Interface for Multiple Source Capture Recapture Models}
\description{
Implements user interfaces for log-linear models, bayesian model averaging and bayesian dirichlet process mixture models.
}
\author{
Ian E. Fellows \email{ian@fellstat.com}
}
